# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "W03 CRUD API",
    "description": "API with two collections (items and categories). Full CRUD, validation, error handling, and GitHub OAuth authentication.",
    "version": "2.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://w03-crud-api-vmrb.onrender.com",
      "description": "Production server on Render (Original)"
    },
    {
      "url": "https://w03-crud-api-1.onrender.com",
      "description": "Production server on Render (Instance 2 - Main)"
    },
    {
      "url": "http://localhost:8083",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "GitHub OAuth2 Authentication - Click 'Authorize' button below and login with GitHub",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://github.com/login/oauth/authorize",
            "tokenUrl": "https://github.com/login/oauth/access_token",
            "scopes": {
              "read:user": "Read user information",
              "user:email": "Read user email addresses"
            }
          }
        }
      },
      "apiKey": {
        "type": "apiKey",
        "name": "API_SECRET",
        "in": "header",
        "description": "API Secret key for additional security"
      }
    },
    "schemas": {
      "Item": {
        "type": "object",
        "required": ["name", "description", "price", "category", "inStock", "sku"],
        "properties": {
          "_id": {
            "type": "string",
            "description": "MongoDB ObjectId"
          },
          "name": {
            "type": "string",
            "description": "Item name",
            "example": "Laptop Computer"
          },
          "description": {
            "type": "string",
            "description": "Item description",
            "example": "High-performance laptop with 16GB RAM"
          },
          "price": {
            "type": "number",
            "description": "Item price",
            "example": 999.99
          },
          "category": {
            "type": "string",
            "description": "Item category",
            "example": "Electronics"
          },
          "inStock": {
            "type": "boolean",
            "description": "Stock availability",
            "example": true
          },
          "sku": {
            "type": "string",
            "description": "Stock Keeping Unit",
            "example": "LAP-001"
          },
          "weight": {
            "type": "number",
            "description": "Item weight in kg",
            "example": 1.5
          },
          "manufacturer": {
            "type": "string",
            "description": "Manufacturer name",
            "example": "TechCorp"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "createdBy": {
            "type": "string",
            "description": "User ID who created the item"
          }
        }
      },
      "Category": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "_id": {
            "type": "string",
            "description": "MongoDB ObjectId"
          },
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Electronics"
          },
          "description": {
            "type": "string",
            "description": "Category description",
            "example": "Electronic devices and accessories"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "createdBy": {
            "type": "string",
            "description": "User ID who created the category"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "oauth2": ["read:user", "user:email"]
    }
  ],
  "paths": {
...
    "/auth/github": {
      "get": {
        "summary": "Authenticate with GitHub",
        "tags": ["Authentication"],
        "description": "Redirects to GitHub for OAuth authentication. Uses Client ID: Ov23liVpFGbPyMJFrYYM",
        "externalDocs": {
          "description": "GitHub OAuth Documentation",
          "url": "https://docs.github.com/en/apps/oauth-apps/building-oauth-apps/authorizing-oauth-apps"
        },
        "responses": {
          "302": {
            "description": "Redirect to GitHub login page",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "GitHub OAuth authorization URL"
              }
            }
          },
          "500": {
            "description": "Authentication error"
          }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "summary": "GitHub OAuth callback",
        "tags": ["Authentication"],
        "description": "GitHub redirects here after authentication. This endpoint should not be called directly.",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Authorization code from GitHub"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to API docs on success or home on failure"
          }
        }
      }
    },
...
    "/items": {
      "get": {
        "summary": "Get all items",
        "tags": ["Items"],
        "description": "Retrieves all items from the database (public access)",
        "security": [],
        "responses": {
          "200": {
            "description": "Array of items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new item",
        "tags": ["Items"],
        "description": "Creates a new item (requires GitHub authentication)",
        "security": [
          {
            "oauth2": ["read:user"]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "description", "price", "category", "inStock", "sku"],
                "properties": {
                  "name": { "type": "string", "example": "Smartphone" },
                  "description": { "type": "string", "example": "Latest model smartphone" },
                  "price": { "type": "number", "example": 699.99 },
                  "category": { "type": "string", "example": "Electronics" },
                  "inStock": { "type": "boolean", "example": true },
                  "sku": { "type": "string", "example": "PHN-001" },
                  "weight": { "type": "number", "example": 0.5 },
                  "manufacturer": { "type": "string", "example": "PhoneCorp" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "id": { "type": "string" },
                    "item": { "$ref": "#/components/schemas/Item" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Please login with GitHub first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "412": {
            "description": "Validation failed – missing required fields or invalid type"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "summary": "Get an item by ID",
        "tags": ["Items"],
        "description": "Retrieves a single item by its MongoDB ObjectId (public access)",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "MongoDB ObjectId (24 hex chars)",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "200": {
            "description": "Item found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an item (partial update allowed)",
        "tags": ["Items"],
        "description": "Updates an existing item (requires GitHub authentication)",
        "security": [
          {
            "oauth2": ["read:user"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number" },
                  "category": { "type": "string" },
                  "inStock": { "type": "boolean" },
                  "sku": { "type": "string" },
                  "weight": { "type": "number" },
                  "manufacturer": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully"
          },
          "400": {
            "description": "Invalid ID format or invalid JSON"
          },
          "401": {
            "description": "Unauthorized - Please login with GitHub first"
          },
          "404": {
            "description": "Item not found"
          },
          "412": {
            "description": "Validation failed for sent fields"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an item",
        "tags": ["Items"],
        "description": "Deletes an item by ID (requires GitHub authentication)",
        "security": [
          {
            "oauth2": ["read:user"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "deletedCount": { "type": "number" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "401": {
            "description": "Unauthorized - Please login with GitHub first"
          },
          "404": {
            "description": "Item not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "summary": "Get all categories",
        "tags": ["Categories"],
        "description": "Retrieves all categories from the database (public access)",
        "security": [],
        "responses": {
          "200": {
            "description": "Array of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new category",
        "tags": ["Categories"],
        "description": "Creates a new category (requires GitHub authentication)",
        "security": [
          {
            "oauth2": ["read:user"]
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "Books" },
                  "description": { "type": "string", "example": "All kinds of books" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Missing 'name' field"
          },
          "401": {
            "description": "Unauthorized - Please login with GitHub first"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "summary": "Get a category by ID",
        "tags": ["Categories"],
        "description": "Retrieves a single category by ID (public access)",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "200": {
            "description": "Category found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Category not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update a category",
        "tags": ["Categories"],
        "description": "Updates an existing category (requires GitHub authentication)",
        "security": [
          {
            "oauth2": ["read:user"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully"
          },
          "400": {
            "description": "Invalid ID or missing name"
          },
          "401": {
            "description": "Unauthorized - Please login with GitHub first"
          },
          "404": {
            "description": "Category not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a category",
        "tags": ["Categories"],
        "description": "Deletes a category by ID (requires GitHub authentication)",
        "security": [
          {
            "oauth2": ["read:user"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully"
          },
          "400": {
            "description": "Invalid ID format"
          },
          "401": {
            "description": "Unauthorized - Please login with GitHub first"
          },
          "404": {
            "description": "Category not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}